backend/pyproject.toml
[tool.poetry]
name = "afishka"
version = "1.0.0"
description = ""
authors = ["AlanTheKnight <51376126+AlanTheKnight@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
Markdown = "^3.4.1"
sorl-thumbnail-serializer-field = "^0.2.1"
toml = "^0.10.2"
django-cors-headers = "^3.13.0"
django-cleanup = "^6.0.0"
Pillow = "^9.2.0"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
